pyproject.toml
[project]
name = "hccda-ocr-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-cognitiveservices-vision-computervision>=0.9.1",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
]
